Remove unneeded files; fix code smells
