--- website/projects/scimaps/src/app/app.component.scss
@@ -1,3 +1,0 @@
-.news-item {
-  width: 31.25rem;
-}
