--- website/projects/scimaps/src/app/pages/maps-page/maps-page.component.html
@@ -1,36 +1,0 @@
-<sci-item-drawer class="details-drawer" [class.hidden]="!showDrawer" (closeDrawer)="closeDetails()"></sci-item-drawer>
-
-<h2 class="header">Maps</h2>
-<div class="highlight-content">
-  <p class="highlight-body">{{  highlightBody }}</p>
-  <div class="carousel-container">
-    <sci-carousel>
-      <sci-carousel-item *ngFor="let slide of slides">
-        <img [src]="slide">
-      </sci-carousel-item>
-    </sci-carousel>
-  </div>
-</div>
-
-<h2 class="header discover-maps">Discover Maps</h2>
-
-<sci-discover-listing class="discover-item" *ngFor="let item of displayItems" [discoverItem]="item" (openDrawer)="openDetails()"></sci-discover-listing>
-
-<button *ngIf="moreMaps()" class="more-button" (click)="showMoreMaps()">Show More Maps</button>
-<div class="mobile-button" (click)="showModal = true;">
-  Discover Maps
-  <mat-icon class="mobile-icon">chevron_right</mat-icon>
-</div>
-
-<div class="fullscreen-modal" *ngIf="showModal">
-  <div class="modal-header">
-    <mat-icon class="mobile-icon" (click)="showModal = false;">chevron_left</mat-icon>
-    <div class="mobile-title"><h3>Maps</h3></div>
-  </div>
-
-  <sci-carousel class="maps-carousel">
-    <sci-carousel-item *ngFor="let discoverItem of discoverItems">
-      <sci-discover-listing [discoverItem]="discoverItem" (openDrawer)="openDetails()"></sci-discover-listing>
-    </sci-carousel-item>
-  </sci-carousel>
-</div>
